requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pandas==2.2.3
numpy==2.1.3
pytrends==4.9.2
